+++ modules/vp8/encode.c
@@ -157,7 +157,7 @@ static inline void hdr_encode(uint8_t hdr[HDR_SIZE], bool noref, bool start,
 
 static inline int packetize(bool marker, const uint8_t *buf, size_t len,
 			    size_t maxlen, bool noref, uint8_t partid,
-			    uint16_t picid, uint32_t rtp_ts,
+			    uint16_t picid, uint64_t rtp_ts,
 			    videnc_packet_h *pkth, void *arg)
 {
 	uint8_t hdr[HDR_SIZE];
@@ -236,7 +236,7 @@ int vp8_encode(struct videnc_state *ves, bool update,
 		bool keyframe = false, marker = true;
 		const vpx_codec_cx_pkt_t *pkt;
 		uint8_t partid = 0;
-		uint32_t ts;
+		uint64_t ts;
 
 		pkt = vpx_codec_get_cx_data(&ves->ctx, &iter);
 		if (!pkt)
