+++ src/vidutil.c
@@ -14,23 +14,23 @@
  * Calculate the RTP timestamp from Presentation Time Stamp (PTS)
  * or Decoding Time Stamp (DTS) and framerate.
  *
- * @note The calculated RTP Timestamp may wrap around.
+ * @note The calculated RTP Timestamp may NOT wrap around.
  *
  * @param pts Presentation Time Stamp (PTS)
  * @param fps Framerate in [frames per second]
  *
- * @return RTP Timestamp
+ * @return Extended RTP Timestamp
  */
-uint32_t video_calc_rtp_timestamp(int64_t pts, unsigned fps)
+uint64_t video_calc_rtp_timestamp(int64_t pts, unsigned fps)
 {
-       uint64_t rtp_ts;
+	uint64_t rtp_ts;
 
-       if (!fps)
-	       return 0;
+	if (!fps)
+		return 0;
 
-       rtp_ts = ((uint64_t)VIDEO_SRATE * pts) / fps;
+	rtp_ts = ((uint64_t)VIDEO_SRATE * pts) / fps;
 
-       return (uint32_t)rtp_ts;
+	return rtp_ts;
 }
